Improve code quality
Currently some of the comments were written following my own intuition
not following the Code Quality standards.

Changes are made to follow the code quality standards, that the first sentence
should start in the form of Returns, Sends, Adds, etc. This change was done to
most of the class header comments as well as some of their constructors