Use Java Streams
Currently some functions including the display function of the tasklist
uses a for loop to iterate through the tasks and append them to a list.

Java however has a cleaner way to implement this using its Stream lib.
We can use it to both filter and concatinate the strings in a clean and
singular command statement.